+++ backend/src/baserow/contrib/database/formula/types/visitors.py
@@ -182,7 +182,9 @@ class FormulaTypingVisitor(
                     )
             # check the lookup field
             expression = field_type.to_baserow_formula_expression(referenced_field)
-            if isinstance(expression, BaserowFunctionCall):
+            # if other formula fields are referenced, we want to avoid
+            # keep nesting wrapper functions, so unwrap the expression here
+            if expression.is_wrapper:
                 expression = expression.expression_type.unwrap_at_field_level(
                     expression
                 )